CCD-2935 Refactor CaseFieldMetadata model
   [CCD-2935](https://tools.hmcts.net/jira/browse/CCD-2935)

Refactor CaseFieldMetadata model.
