+++ src/main/java/uk/gov/hmcts/ccd/domain/service/common/CaseFieldMetadataExtractor.java
@@ -3,8 +3,8 @@ package uk.gov.hmcts.ccd.domain.service.common;
 import com.fasterxml.jackson.databind.JsonNode;
 import io.vavr.control.Either;
 import lombok.Value;
+import uk.gov.hmcts.ccd.domain.model.common.CaseFieldMetadata;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseFieldDefinition;
-import uk.gov.hmcts.ccd.domain.model.definition.CaseFieldMetadata;
 import uk.gov.hmcts.ccd.domain.types.BaseType;
 
 import java.util.List;
